Move functions to a modular class
