+++ whack_a_mole/whack_a_mole/camera.py
@@ -12,17 +12,8 @@ from sensor_msgs.msg import Image, CameraInfo
 
 from tf2_ros import TransformBroadcaster
 from geometry_msgs.msg import TransformStamped
-
-global COLORS, COLORS_HSV
-
-COLORS = {"GREEN": 0, "YELLOW": 1, "BLUE": 2, "RED": 3}
-
-COLORS_HSV = {
-    "GREEN": [np.array([47, 90, 132]), np.array([94, 255, 191])],
-    "YELLOW": [np.array([24, 100, 101]), np.array([52, 255, 230])],
-    "BLUE": [np.array((95, 157, 85)), np.array((161, 255, 189))],
-    "RED": [np.array((0, 132, 4)), np.array((9, 241, 221))],
-}
+from whack_a_mole.mole_detection import OpenCVClient
+from whack_a_mole.constants import COLORS, COLORS_HSV
 
 
 class Camera(Node):
@@ -70,12 +61,12 @@ class Camera(Node):
         )
         
         self.create_service(Empty, 'toggle_tf_publish', self.toggle_tf_publish)
-
+        self.clipping_distance = self.get_parameter("clipping_distance").value
+        self.cv2_client = OpenCVClient(
+            clipping_distance=self.clipping_distance,
+            crop=self.crop
+        )
         self.tf_broadcaster = TransformBroadcaster(self, qos=QoSProfile(depth=10))
-
-        self.color_image = np.array([])
-        self.depth_image = np.array([])
-        self.camera_intrinsics = np.array([])
 
         self.prev_xyz = np.array([0, 0, 0], dtype=float)
 
@@ -88,13 +79,11 @@ class Camera(Node):
             + (np.array(self.crop[0]) + np.array(self.crop[1])) / 2
         )
 
-        self.clipping_distance = 1400
-        
         self.update_colors = True
 
     def timer_callback(self):
-
-        if self.color_image.shape[0] == 0 or self.depth_image.shape[0] == 0:
+        """Timer callback to detect the color of the object and broadcast the frame."""
+        if self.cv2_client.color_image.shape[0] == 0 or self.cv2_client.depth_image.shape[0] == 0:
             return
 
         for color in COLORS_HSV.keys():
@@ -108,19 +97,35 @@ class Camera(Node):
             self.detect_illumination(color)
 
     def toggle_tf_publish(self, request, response):
-        
         self.update_colors = not self.update_colors
         return response
 
     def detect_illumination(self, color: str):
-        self.log("Detecting illumination for ", color)
-        self.log(f"{self.running_avg}")
+        """
+        Detects the illumination of the color and broadcasts.
+        
+        Args:
+            color (str): The color of the object
+        
+        """
+        color_index = COLORS[color]
+        median_centroid = np.median(self.running_avg[color_index], axis=0)
+        median_centroid = np.array(median_centroid, dtype=int)
 
     def broadcast_color(self, lower_HSV, higher_HSV, color: str):
-
+        """
+        Broadcasts the color frame.
+        
+        Args:
+            lower_HSV (np.array): The lower HSV range
+            higher_HSV (np.array): The higher HSV range
+            color (str): The color of the object
+        """
         color_index = COLORS[color]
 
-        x_c, y_c = self.find_color_centroid(lower_HSV, higher_HSV)
+        x_c, y_c, bg_removed = self.find_color_centroid(lower_HSV, higher_HSV)
+        msg = CvBridge().cv2_to_imgmsg(bg_removed, encoding="bgr8")
+        self.image_pub.publish(msg)
 
         if x_c != 0 and y_c != 0 and self.update_colors:
 
@@ -137,147 +142,35 @@ class Camera(Node):
         )
 
     def log(self, *message):
-
+        """Custom logger function."""
         self.get_logger().info(f"[CAMERA NODE]: {", ".join(str(i) for i in message)}")
 
     def camera_info_callback(self, msg):
-        # Save the camera intrinsics
-        self.camera_intrinsics = msg.k
-
-    def get_depth_info(self, msg: Image):
-
-        self.depth_image = CvBridge().imgmsg_to_cv2(msg)
-
-    def get_color_info(self, msg: Image):
-
-        self.color_image = CvBridge().imgmsg_to_cv2(msg, desired_encoding="bgr8")
-
-    def find_color_centroid(self, lower_HSV: np.array, higher_HSV: np.array):
         """
-        Returns the pixel indecies of the centroid of a color defined in lower_HSV , higher_HSV range
-        """
-
-        depth_image_3d = np.dstack(
-            (self.depth_image, self.depth_image, self.depth_image)
-        )  # depth image is 1 channel, color is 3 channels
-        bg_removed = np.where(
-            (depth_image_3d > self.clipping_distance) | (depth_image_3d <= 0),
-            0,
-            self.color_image,
-        )
-
-        if lower_HSV.shape[0] == 2:
-            mask1 = cv2.inRange(
-                cv2.cvtColor(bg_removed, cv2.COLOR_BGR2HSV), lower_HSV[0], higher_HSV[0]
-            )
-            mask2 = cv2.inRange(
-                cv2.cvtColor(bg_removed, cv2.COLOR_BGR2HSV), lower_HSV[1], higher_HSV[1]
-            )
-            mask = mask1 + mask2
-
-        else:
-
-            mask = cv2.inRange(
-                cv2.cvtColor(bg_removed, cv2.COLOR_BGR2HSV), lower_HSV, higher_HSV
-            )
-
-        if not np.any(mask):
-            return np.array([-1, -1])
-
-        # Find Centroid
-        x_c, y_c = self.find_centroid_cropped(mask, (*self.crop[0], *self.crop[1]))
-
-        x_c = int(x_c)
-        y_c = int(y_c)
-
-        for color in COLORS:
-            color_index = COLORS[color]
-
-            avg_centroid = np.median(self.running_avg[color_index], axis=0)
-            avg_centroid = np.array(avg_centroid, dtype=int)
-
-            cv2.circle(
-                bg_removed,
-                (avg_centroid[0], avg_centroid[1]),
-                5,
-                (0, 0, 255),
-                thickness=10,
-            )
-
-            cv2.putText(
-                bg_removed,
-                f"{color}",
-                (avg_centroid[0], avg_centroid[1]),
-                cv2.FONT_HERSHEY_SIMPLEX,
-                1,
-                (255, 255, 255),
-                2,
-            )
-
-        cv2.rectangle(bg_removed, self.crop[0], self.crop[1], (0, 0, 255), thickness=5)
-
-        msg = CvBridge().cv2_to_imgmsg(bg_removed, encoding="bgr8")
-
-        self.image_pub.publish(msg)
-
-        return x_c, y_c
-
-    def find_centroid_cropped(self, masked_image, crop_indices):
-        """
-        Find the centroid of the largest contour within a specified crop region.
+        Callback for camera info subscriber
 
         Args:
-        - masked_image (np.array): The masked image to process.
-        - crop_indices (tuple): A tuple (x1, y1, x2, y2) specifying the crop region.
-
-        Returns:
-        - (int, int): The (x, y) coordinates of the centroid in the original image.
+            msg (CameraInfo): Camera Info message
         """
+        self.cv2_client.camera_intrinsics = msg.k
 
-        # Unpack the crop region indices
-        x1, y1, x2, y2 = crop_indices
+    def get_depth_info(self, msg: Image):
+        """
+        Callback for depth image subscriber
+        
+        Args:
+            msg (Image): Depth image message
+        """
+        self.cv2_client.depth_image = CvBridge().imgmsg_to_cv2(msg)
 
-        # Crop the masked image based on the provided indices
-        cropped_image = masked_image[y1:y2, x1:x2]
-
-        # Convert cropped image to numpy array (if necessary)
-        image = np.array(cropped_image, dtype=np.uint8)
-
-        # Find all contours in the cropped region
-        contours, _ = cv2.findContours(image, cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE)
-
-        # Find the largest contour area
-        largest_area = -1
-        largest_index = -1
-
-        for i in range(len(contours)):
-            area = cv2.contourArea(contours[i])
-
-            # Check if the current contour has the largest area so far
-            if area > largest_area:
-                largest_area = area
-                largest_index = i
-
-        # If no valid contour was found (largest_area remains 0)
-        if (largest_area == 0) or (largest_area == -1):
-            return 0, 0
-
-        # Get the centroid in the cropped region
-
-        moments = cv2.moments(contours[largest_index])
-
-        if moments["m00"] == 0:  # To avoid division by zero
-            return 0, 0
-
-        # Calculate the centroid in the cropped image
-        x_c = int(moments["m10"] / moments["m00"])
-        y_c = int(moments["m01"] / moments["m00"])
-
-        # Convert the cropped image centroid back to original image coordinates
-        original_x_c = x1 + x_c
-        original_y_c = y1 + y_c
-
-        return original_x_c, original_y_c
+    def get_color_info(self, msg: Image):
+        """
+        Callback for color image subscriber
+        
+        Args:
+            msg (Image): Color image message
+        """
+        self.cv2_client.color_image = CvBridge().imgmsg_to_cv2(msg, desired_encoding="bgr8")
 
     def broadcast_color_frame(self, base_frame, child_frame, x_c, y_c):
 
@@ -305,35 +198,6 @@ class Camera(Node):
         transform.transform.rotation.w = 1.0
 
         self.tf_broadcaster.sendTransform(transform)
-
-    def get_3d_coordinates_at_pixel(self, x, y, frame_name=""):
-        """Convert the pixel coordinates (x, y) to 3D camera coordinates in meters."""
-
-        if self.camera_intrinsics.shape[0] == 0:
-            self.log("Camera intrinsics not yet received.")
-            return -1, -1, -1
-
-        # Get the depth value at the specified pixel (x, y)
-        depth_in_meters = self.depth_image[y, x] / 1000
-
-        if depth_in_meters == 0:  # No depth data at this pixel
-            return -1, -1, -1
-
-        # Extract intrinsic parameters from the camera info (in a 3x3 matrix form)
-        f_x, f_y = (
-            self.camera_intrinsics[0],
-            self.camera_intrinsics[4],
-        )  # Focal lengths (in pixels)
-        c_x, c_y = (
-            self.camera_intrinsics[2],
-            self.camera_intrinsics[5],
-        )  # Optical center (in pixels)
-
-        # Convert to 3D coordinates (X, Y, Z) in camera frame
-        X = (x - c_x) * depth_in_meters / f_x
-        Y = (y - c_y) * depth_in_meters / f_y
-        Z = depth_in_meters  # Depth in meters
-        return X, Y, Z
 
 
 def entry(args=None):
