+++ whack_a_mole/whack_a_mole/constants.py
@@ -1,0 +1,9 @@
+import numpy as np
+COLORS = {"GREEN": 0, "YELLOW": 1, "BLUE": 2, "RED": 3}
+
+COLORS_HSV = {
+    "GREEN": [np.array([47, 90, 132]), np.array([94, 255, 191])],
+    "YELLOW": [np.array([24, 100, 101]), np.array([52, 255, 230])],
+    "BLUE": [np.array((95, 157, 85)), np.array((161, 255, 189))],
+    "RED": [np.array((0, 132, 4)), np.array((9, 241, 221))],
+}
